+++ Pandora/Paradigm/Structure/Some/Splay.hs
@@ -5,20 +5,21 @@ module Pandora.Paradigm.Structure.Some.Splay where
 import Pandora.Core.Functor (type (~>), type (:.), type (>))
 import Pandora.Pattern.Semigroupoid ((.))
 import Pandora.Pattern.Category ((<--), (<---), (<----), (<-----), (<------), identity)
-import Pandora.Pattern.Functor.Covariant (Covariant ((<-|-)))
+import Pandora.Pattern.Functor.Covariant (Covariant ((<-|-), (<-|--)))
 import Pandora.Pattern.Functor.Bindable (Bindable ((==<<), (===<<)))
 import Pandora.Pattern.Transformer.Hoistable ((/|\))
 import Pandora.Paradigm.Primary ()
-import Pandora.Paradigm.Primary.Algebraic ((.:..), (<-*-), extract)
+import Pandora.Paradigm.Primary.Algebraic ((<-*-), extract)
+import Pandora.Paradigm.Primary.Algebraic.Product (type (<:*:>))
 import Pandora.Paradigm.Primary.Functor.Maybe (Maybe (Just))
 import Pandora.Paradigm.Primary.Functor.Tagged (type (:#))
 import Pandora.Paradigm.Primary.Functor.Wye (Wye (Left, Right))
 import Pandora.Paradigm.Primary.Transformer.Construction (Construction (Construct), deconstruct)
 import Pandora.Paradigm.Primary (twosome)
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (run)
-import Pandora.Paradigm.Inventory.Some.Optics (mutate)
+import Pandora.Paradigm.Inventory.Some.Optics (view, mutate)
 import Pandora.Paradigm.Schemes (TT (TT), type (<::>))
-import Pandora.Paradigm.Structure.Ability.Morphable (Morphable (Morphing, morphing), Morphed, Morph (Rotate, Into), premorph, rotate, into)
+import Pandora.Paradigm.Structure.Ability.Morphable (Morphable (Morphing, morphing), Morphed, Morph (Rotate), premorph, rotate)
 import Pandora.Paradigm.Structure.Ability.Nonempty (Nonempty)
 import Pandora.Paradigm.Structure.Ability.Substructure (sub)
 import Pandora.Paradigm.Structure.Ability.Monotonic (resolve)
@@ -52,60 +53,58 @@ instance Morphable (Rotate > Right > Zig Zag) Binary where
 
 -------------------------------------- Non-empty Splay tree ----------------------------------------
 
-instance Morphable (Rotate > Left Zig) (Construction Wye) where
-	type Morphing (Rotate > Left Zig) (Construction Wye) = Binary
-	morphing :: forall a . (:#) (Rotate > Left Zig) <::> Construction Wye > a -> Binary a
+-- TODO: refactor so that there is only one expression
+instance Morphable (Rotate > Left Zig) (Construction (Maybe <:*:> Maybe)) where
+	type Morphing (Rotate > Left Zig) (Construction (Maybe <:*:> Maybe)) = Binary
+	morphing :: forall a . (:#) (Rotate > Left Zig) <::> Construction (Maybe <:*:> Maybe) > a -> Binary a
 	morphing (premorph -> Construct x xs) = TT <--- Construct <-|- parent <-*- Just nodes where
 
-		nodes :: Wye :. Nonempty Binary > a
-		nodes = into @Wye .:.. twosome
-			<------ branch @Left xs
-			<------ Just . Construct x . into @Wye
+		nodes :: (Maybe <:*:> Maybe) :. Nonempty Binary > a
+		nodes = twosome
+			<------ view <-- sub @Left <-- xs
+			<------ Just . Construct x
 				<----- twosome
-					<---- branch @Left ===<< deconstruct <-|- branch @Right xs
-					<---- branch @Right ===<< deconstruct <-|- branch @Right xs
+					<---- view (sub @Left) ===<< deconstruct <-|- view (sub @Right) xs
+					<---- view (sub @Right) ===<< deconstruct <-|- view (sub @Right) xs
 
 		parent :: Maybe a
-		parent = extract <-|- branch @Right xs
+		parent = extract <-|-- view <-- sub @Right <-- xs
 
-instance Morphable (Rotate > Right Zig) (Construction Wye) where
-	type Morphing (Rotate > Right Zig) (Construction Wye) = Binary
-	morphing :: forall a . (:#) (Rotate > Right Zig) <::> Construction Wye > a -> Binary a
+instance Morphable (Rotate > Right Zig) (Construction (Maybe <:*:> Maybe)) where
+	type Morphing (Rotate > Right Zig) (Construction (Maybe <:*:> Maybe)) = Binary
+	morphing :: forall a . (:#) (Rotate > Right Zig) <::> Construction (Maybe <:*:> Maybe) > a -> Binary a
 	morphing (premorph -> Construct x xs) = TT <--- Construct <-|- parent <-*- Just nodes where
 
-		nodes :: Wye :. Nonempty Binary > a
-		nodes = into @Wye .:.. twosome
-			<------ branch @Left ===<< deconstruct <-|- branch @Left xs
-			<------ Just . Construct x . into @Wye
+		nodes :: (Maybe <:*:> Maybe) :. Nonempty Binary > a
+		nodes = twosome
+			<------ view (sub @Left) ===<< deconstruct <-|- view (sub @Left) xs
+			<------ Just . Construct x
 				<----- twosome
-					<---- branch @Right ===<< deconstruct <-|- branch @Left xs
-					<---- branch @Right xs
+					<---- view (sub @Right) ===<< deconstruct <-|- view (sub @Left) xs
+					<---- view (sub @Right) xs
 
 		parent :: Maybe a
-		parent = extract <-|- branch @Left xs
+		parent = extract <-|-- view <-- sub @Left <-- xs
 
 -- TODO: Morphing ... = Conclussion Error <::> Nonempty Binary
-instance Morphable (Rotate > Left > Zig Zig) (Construction Wye) where
-	type Morphing (Rotate > Left > Zig Zig) (Construction Wye) = Maybe <::> Construction Wye
+instance Morphable (Rotate > Left > Zig Zig) (Construction (Maybe <:*:> Maybe)) where
+	type Morphing (Rotate > Left > Zig Zig) (Construction (Maybe <:*:> Maybe)) = Maybe <::> Construction (Maybe <:*:> Maybe)
 	morphing (premorph -> tree) = TT <---- run . rotate @(Left Zig) ===<< run <-- rotate @(Left Zig) tree
 
 -- TODO: Morphing ... = Conclussion Error <::> Nonempty Binary
-instance Morphable (Rotate > Right > Zig Zig) (Construction Wye) where
-	type Morphing (Rotate > Right > Zig Zig) (Construction Wye) = Maybe <::> Construction Wye
+instance Morphable (Rotate > Right > Zig Zig) (Construction (Maybe <:*:> Maybe)) where
+	type Morphing (Rotate > Right > Zig Zig) (Construction (Maybe <:*:> Maybe)) = Maybe <::> Construction (Maybe <:*:> Maybe)
 	morphing (premorph -> tree) = TT <---- run . rotate @(Right Zig) ===<< run <-- rotate @(Right Zig) tree
 
 -- TODO: Morphing ... = Conclussion Error <::> Nonempty Binary
-instance Morphable (Rotate > Left > Zig Zag) (Construction Wye) where
-	type Morphing (Rotate > Left > Zig Zag) (Construction Wye) = Maybe <::> Construction Wye
+instance Morphable (Rotate > Left > Zig Zag) (Construction (Maybe <:*:> Maybe)) where
+	type Morphing (Rotate > Left > Zig Zag) (Construction (Maybe <:*:> Maybe)) = Maybe <::> Construction (Maybe <:*:> Maybe)
 	morphing (premorph -> struct) = rotate @(Left Zig) <--- mutate <-- (try_to_rotate @(Right Zig) /|\) <-- sub @Left <-- struct
 
 -- TODO: Morphing ... = Conclussion Error <::> Nonempty Binary
-instance Morphable (Rotate > Right > Zig Zag) (Construction Wye) where
-	type Morphing (Rotate > Right > Zig Zag) (Construction Wye) = Maybe <::> Construction Wye
+instance Morphable (Rotate > Right > Zig Zag) (Construction (Maybe <:*:> Maybe)) where
+	type Morphing (Rotate > Right > Zig Zag) (Construction (Maybe <:*:> Maybe)) = Maybe <::> Construction (Maybe <:*:> Maybe)
 	morphing (premorph -> struct) = rotate @(Right Zig) <--- mutate <-- (try_to_rotate @(Left Zig) /|\) <-- sub @Right <-- struct
-
-branch :: forall b . Morphable (Into > b Maybe) Wye => Wye ~> Morphing (Into > b Maybe) Wye
-branch = into @(b Maybe)
 
 -- TODO: Include error instead of returning empty tree
 try_to_rotate :: forall direction . Morphed (Rotate direction) (Nonempty Binary) Binary => Nonempty Binary ~> Nonempty Binary